backend/requirements.txt
# Core web stack
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.3

# Optional (heavy) ML dependencies. The app degrades gracefully if these are
# missing — transcription falls back to a placeholder and NER uses a
# lightweight rule-based engine.
#
# Whisper (self-hosted speech-to-text)
#   openai-whisper (pulls in torch). On Python 3.14 you may prefer faster-whisper
#   if torch wheels aren't yet published. Install either:
#       pip install openai-whisper
#       pip install faster-whisper ctranslate2
#
# Turkish morphology & NER:
zeyrek>=0.1.3
# Robust Turkish relative/absolute date grounding (geçen hafta bugün, 14 Nisan...):
dateparser>=1.2
# Local LLM for structured event extraction (Qwen2.5-7B GGUF). The
# pipeline falls back to NLP-only output when the model file is absent.
llama-cpp-python>=0.3
# transformers + torch for BERTurk NER (optional, large download):
#   pip install transformers torch

# Production server (optional)
gunicorn>=22.0
